Introduce before and after steps for a task. (#56)
* Introduce before and after steps for a task.

A task is now able to have several steps which are executed before and after its own execution. The config syntax for this is the same as for the task itself: 'command', 'script' and 'exec' are supported. A task and all of the preceeding steps will still be executed even when the before steps are failing.

Running a task now returns a list of errors which are logged according to their occurrence for easier troubleshooting.

Co-authored-by: Jens Petersohn <me@jen.pet>